requirements.txt
openai
huggingface_hub
python-dotenv
numpy
requests
google-generativeai
